requirements.txt
pyttsx3==2.90
gtts==2.4.0
python-dotenv==1.0.0
deepgram-sdk==3.2.7
nltk==3.8.1
pydub==0.25.1
moviepy==1.0.3
playsound==1.3.0
